refactor: moving EditProductDTO to dtos/products
