+++ backend/zerowaste/src/main/java/com/zerowaste/services/products/ProductService.java
@@ -7,7 +7,7 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.zerowaste.dtos.EditProductDTO;
+import com.zerowaste.dtos.products.EditProductDTO;
 import com.zerowaste.models.product.Product;
 import com.zerowaste.models.product.ProductCategory;
 import com.zerowaste.repositories.ProductsRepository;
